Added tagPartSplitter parameter for tag sorting
If you use a tag format like v_1.2.3.4 then lexicographical order does
not sort properly because a given version component could have different
numbers of digits.  To fix this sorting can be done with awareness of
the tokens that make up the tag. To break the tag into tokens you
specify a tagPartSplitter pattern that acts as a regular expression. If
the parameter is blank then a direct lexicographical comparison will be
used.