Changed tag map to make sorting work correctly
In the last commit I already implemented tag sorting with the tag part
splitter parameter, but it still wasn't actually showing up in the right
order in jenkins. The collection of tags is stored in a java map where
previously both the key and value were the name of the tag which caused
the order to get lost even though the tags were being inserted to the
map in the right order. Now the keys are just integers starting from 0
and the values are the tag names. It appears that jenkins sorts by tag
name because it seems to be showing up correctly now.